Reorganise lib directory.
Also make the loader code use Railties now.
